1. Tradutor_Artigos-Tecnicos_AzureAI/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asttokens==2.4.1
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
comm==0.2.2
debugpy==1.8.9
decorator==5.1.1
distro==1.9.0
docx==0.2.4
executing==2.1.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
ipykernel==6.29.5
ipython==8.29.0
jedi==0.19.2
jiter==0.8.0
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.7.2
langchain-core==0.3.21
langchain-openai==0.2.10
langsmith==0.1.147
lxml==5.3.0
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
openai==1.55.2
orjson==3.10.12
packaging==24.2
parso==0.8.4
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.1.0
pure_eval==0.2.3
pydantic==2.10.2
pydantic_core==2.27.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
setuptools==75.6.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
tenacity==9.0.0
tiktoken==0.8.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.2.3
wcwidth==0.2.13